Major update to Hal Components
