+++ app/src/index.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
 const container = document.getElementById("root");
-const root = createRoot(container);
+const root = createRoot(container!);
 root.render(<App />);
 
 // If you want your app to work offline and load faster, you can change
