--- lib/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["@babel/react", "@babel/env"],
-  "plugins": ["@babel/plugin-proposal-class-properties", "@babel/plugin-transform-runtime"]
-}
