+++ lib/package.json
@@ -9,22 +9,21 @@
     "url": "https://dxc.com"
   },
   "main": "./dist/main.js",
+  "types": "./dist/main.d.ts",
   "license": "Apache-2.0",
   "peerDependencies": {
     "react": "^18.x",
     "react-dom": "^18.x",
-    "@dxc-technology/halstack-react": "6.x || 0.0.0-x"
+    "@dxc-technology/halstack-react": "8.0.0"
   },
   "dependencies": {
-    "@dxc-technology/halstack-client": "^1.3.1",
-    "prop-types": "^15.7.2",
-    "react-scripts": "5.0.1"
+    "@dxc-technology/halstack-client": "^1.3.3"
   },
   "scripts": {
     "test": "jest",
     "test:watch": "npm test -- --watch --coverage",
-    "build": "babel src --out-dir dist --copy-files --verbose && node ../scripts/build/copy-readme.js",
-    "build:watch": "babel src --watch --out-dir dist --copy-files --verbose"
+    "build": "babel src --extensions .js,.jsx,.ts,.tsx --out-dir dist --copy-files --verbose && node ../scripts/build/copy-package.js && tsc ",
+    "build:watch": "babel src --watch --extensions .js,.jsx,.ts,.tsx --out-dir dist --copy-files --verbose"
   },
   "devDependencies": {
     "@babel/cli": "^7.6.2",
@@ -33,6 +32,10 @@
     "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/preset-env": "^7.6.2",
     "@babel/preset-react": "^7.0.0",
+    "@dxc-technology/halstack-react": "8.0.0",
+    "@types/react": "^18.0.27",
+    "@types/styled-components": "^5.1.26",
+    "@types/jest": "^29.4.0",
     "babel-jest": "^29.3.1",
     "babel-loader": "^9.1.0",
     "eslint": "^5.16.0",
@@ -42,6 +45,9 @@
     "eslint-plugin-jest": "^22.7.1",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.13.0",
-    "jest": "^29.2.0"
+    "jest": "^29.2.0",
+    "styled-components": "^5.0.1",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5"
   }
 }
